+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/).
+## Unreleased
+
+### Fixed
+- Proper ordering of handlers and processors.
+- `ContainerInteropLoggerFactory` resolves logger configuration from either `Config` and `config` container entries.
+
+## 0.1.0 - 2017-06-11